requirements.txt
python-barcode==0.15.1
pandas
pillow
pillow
# reportlab
